Refactor error handling logic from withErrorHandler to custom hook
